requirements.txt
Flask
tensorflow
numpy
scikit-learn
Pillow
